small additions

+++ src/Aufgabe5/formicarium/Part.java
@@ -2,6 +2,6 @@ package formicarium;
 
 public interface Part extends Rated<Part, Quality> {
 
-    Numeric value();
+    double value();
     String toString();
 }
